Add AvroGenHelper
